separate configuration as a tester from running all the Ceph s3 tests
