+++ src/test/system_tests/ceph_s3_tests/s3-tests-lists/s3_tests_pending_list.txt
@@ -1,0 +1,93 @@
+s3tests_boto3.functional.test_s3.test_get_bucket_policy_status
+s3tests.functional.test_s3.test_atomic_dual_conditional_write_1mb
+s3tests_boto3.functional.test_s3.test_bucket_listv2_encoding_basic
+s3tests_boto3.functional.test_s3.test_bucket_list_encoding_basic
+s3tests_boto3.functional.test_s3.test_bucket_list_delimiter_none
+s3tests_boto3.functional.test_s3.test_bucket_listv2_delimiter_none
+s3tests_boto3.functional.test_s3.test_bucket_list_delimiter_not_skip_special
+s3tests_boto3.functional.test_s3.test_bucket_list_prefix_basic
+s3tests_boto3.functional.test_s3.test_bucket_list_prefix_delimiter_basic
+s3tests_boto3.functional.test_s3.test_account_usage
+s3tests_boto3.functional.test_s3.test_head_bucket_usage
+s3tests_boto3.functional.test_s3.test_multi_object_delete_key_limit
+s3tests_boto3.functional.test_s3.test_post_object_invalid_signature
+s3tests_boto3.functional.test_s3.test_post_object_invalid_access_key
+s3tests_boto3.functional.test_s3.test_post_object_missing_policy_condition
+s3tests_boto3.functional.test_s3.test_post_object_request_missing_policy_specified_field
+s3tests_boto3.functional.test_s3.test_post_object_expired_policy
+s3tests_boto3.functional.test_s3.test_post_object_invalid_request_field_value
+s3tests_boto3.functional.test_s3.test_bucket_create_exists
+s3tests_boto3.functional.test_s3.test_bucket_create_exists_nonowner
+s3tests_boto3.functional.test_s3.test_bucket_recreate_overwrite_acl
+s3tests_boto3.functional.test_s3.test_bucket_recreate_new_acl
+s3tests_boto3.functional.test_s3.test_list_multipart_upload_owner
+s3tests_boto3.functional.test_s3.test_lifecycle_set
+s3tests_boto3.functional.test_s3.test_lifecycle_expiration
+s3tests_boto3.functional.test_s3.test_lifecycle_expiration_versioning_enabled
+s3tests_boto3.functional.test_s3.test_lifecycle_expiration_tags2
+s3tests_boto3.functional.test_s3.test_lifecycle_expiration_versioned_tags2
+s3tests_boto3.functional.test_s3.test_lifecycle_expiration_noncur_tags1
+s3tests_boto3.functional.test_s3.test_lifecycle_set_date
+s3tests_boto3.functional.test_s3.test_lifecycle_expiration_header_put
+s3tests_boto3.functional.test_s3.test_lifecycle_expiration_header_head
+s3tests_boto3.functional.test_s3.test_lifecycle_expiration_header_tags_head
+s3tests_boto3.functional.test_s3.test_put_obj_enc_conflict_c_s3
+s3tests_boto3.functional.test_s3.test_put_obj_enc_conflict_c_kms
+s3tests_boto3.functional.test_s3.test_put_obj_enc_conflict_s3_kms
+s3tests_boto3.functional.test_s3.test_put_obj_enc_conflict_bad_enc_kms
+s3tests_boto3.functional.test_s3.test_bucket_policy_put_obj_s3_noenc
+s3tests_boto3.functional.test_s3.test_bucket_policy_put_obj_s3_kms
+s3tests_boto3.functional.test_s3.test_bucket_policy_put_obj_kms_noenc
+s3tests_boto3.functional.test_s3.test_bucket_policy_put_obj_kms_s3
+s3tests_boto3.functional.test_s3.test_object_lock_multi_delete_object_with_retention
+s3tests_boto3.functional.test_s3.test_object_lock_put_legal_hold
+s3tests_boto3.functional.test_s3.test_object_lock_get_legal_hold
+s3tests_boto3.functional.test_s3.test_object_lock_changing_mode_from_governance_with_bypass
+s3tests_boto3.functional.test_s3.test_object_lock_changing_mode_from_governance_without_bypass
+s3tests_boto3.functional.test_s3.test_object_lock_changing_mode_from_compliance
+s3tests_boto3.functional.test_s3.test_object_read_unreadable
+s3tests_boto3.functional.test_s3.test_user_policy
+s3tests_boto3.functional.test_s3.test_get_public_acl_bucket_policy_status
+s3tests_boto3.functional.test_s3.test_get_authpublic_acl_bucket_policy_status
+s3tests_boto3.functional.test_s3.test_get_publicpolicy_acl_bucket_policy_status
+s3tests_boto3.functional.test_s3.test_get_nonpublicpolicy_acl_bucket_policy_status
+s3tests_boto3.functional.test_s3.test_get_nonpublicpolicy_deny_bucket_policy_status
+s3tests_boto3.functional.test_s3.test_get_default_public_block
+s3tests_boto3.functional.test_s3.test_put_public_block
+s3tests_boto3.functional.test_s3.test_block_public_put_bucket_acls
+s3tests_boto3.functional.test_s3.test_block_public_object_canned_acls
+s3tests_boto3.functional.test_s3.test_block_public_policy
+s3tests_boto3.functional.test_s3.test_ignore_public_acls
+s3tests_boto3.functional.test_s3.test_sse_s3_default_post_object_authenticated_request
+s3tests_boto3.functional.test_s3.test_sse_kms_default_post_object_authenticated_request
+s3tests_boto3.functional.test_s3select.test_generate_where_clause
+s3tests_boto3.functional.test_s3select.test_generate_projection
+s3tests_boto3.functional.test_s3select.test_count_operation
+s3tests_boto3.functional.test_s3select.test_column_sum_min_max
+s3tests_boto3.functional.test_s3select.test_nullif_expressions
+s3tests_boto3.functional.test_s3select.test_nulliftrue_expressions
+s3tests_boto3.functional.test_s3select.test_is_not_null_expressions
+s3tests_boto3.functional.test_s3select.test_lowerupper_expressions
+s3tests_boto3.functional.test_s3select.test_in_expressions
+s3tests_boto3.functional.test_s3select.test_true_false_in_expressions
+s3tests_boto3.functional.test_s3select.test_like_expressions
+s3tests_boto3.functional.test_s3select.test_truefalselike_expressions
+s3tests_boto3.functional.test_s3select.test_complex_expressions
+s3tests_boto3.functional.test_s3select.test_alias
+s3tests_boto3.functional.test_s3select.test_alias_cyclic_refernce
+s3tests_boto3.functional.test_s3select.test_datetime
+s3tests_boto3.functional.test_s3select.test_true_false_datetime
+s3tests_boto3.functional.test_s3select.test_csv_parser
+s3tests_boto3.functional.test_s3select.test_csv_definition
+s3tests_boto3.functional.test_s3select.test_schema_definition
+s3tests_boto3.functional.test_s3select.test_when_then_else_expressions
+s3tests_boto3.functional.test_s3select.test_coalesce_expressions
+s3tests_boto3.functional.test_s3select.test_cast_expressions
+s3tests_boto3.functional.test_s3select.test_version
+s3tests_boto3.functional.test_s3select.test_trim_expressions
+s3tests_boto3.functional.test_s3select.test_truefalse_trim_expressions
+s3tests_boto3.functional.test_s3select.test_escape_expressions
+s3tests_boto3.functional.test_s3select.test_case_value_expressions
+s3tests_boto3.functional.test_s3select.test_bool_cast_expressions
+s3tests_boto3.functional.test_s3select.test_progress_expressions
+s3tests_boto3.functional.test_s3select.test_output_serial_expressions
